fix(security): resolve 5 high-severity concurrency and memory-safety bugs
- rate_limiter: run cleanup before binding the map reference in
  allow_request, fixing a use-after-free when a 24h-stale entry was
  erased while still referenced
- key_rotation: extract rotate_unlocked so emergency_revoke no longer
  re-locks the non-recursive keys_mutex_ (self-deadlock); make
  approve_rotation take keys_mutex_ before pending_mutex_ to remove the
  AB-BA deadlock with the rotation path
- transport: compute aad_n + ct_n in size_t in parse_frame so the DATA
  length check cannot be bypassed by uint32_t wraparound (OOB read)
- handshake: derive two directional session keys assigned by role so the
  initiator tx key matches the responder rx key, and vice versa

+++ src/transport.hpp
@@ -105,7 +105,7 @@ inline void write_u64(Bytes& out, uint64_t v){ for(int i=0;i<8;i++) out.push_bac
             f.negotiate = NegotiatePayload{parse_features(fb)}; break;
         }
         case FrameType::DATA: {
-            need(4+4); uint32_t aad_n=read_u32(&b[off]); off+=4; uint32_t ct_n=read_u32(&b[off]); off+=4; need(aad_n+ct_n);
+            need(4+4); uint32_t aad_n=read_u32(&b[off]); off+=4; uint32_t ct_n=read_u32(&b[off]); off+=4; need(static_cast<size_t>(aad_n)+static_cast<size_t>(ct_n));
             DataPayload dp{}; dp.aad.assign(b.begin()+off, b.begin()+off+aad_n); off+=aad_n; dp.ciphertext.assign(b.begin()+off, b.begin()+off+ct_n); off+=ct_n;
             f.data = dp; break;
         }
